feat!: Mirror tokio Stream Api with split and into_split
VsockStream has been reworked to have a Owned (the previous API) and a
Borrowed (the new API) variant. This comes with the following
**Breaking Changes**:

- `VsockStream::split(self)` has been renamed `VsockStream::into_split`
- `VsockStream::split(&mut self)` has been added
- `split::{ReadHalf, WriteHalf}` have been renamed to
  `split::{OwnedReadHalf, OwnedWriteHalf}`
- New structs have been added:
  - `split::ReadHalf<'a>` a borrowed ReadHalf of a VsockStream
  - `split::WriteHalf<'a>` a borrowed WriteHalf of a VsockStream

The borrowed variants have the benefit of beeing chaper to
construct/use as they rely on the fact that a `&mut` must be unique,
thus when they are dropped the VsockStream ca be safely used again.

It also helps making it consistant with the tokio sockets.

Signed-off-by: Jalil David Salamé Messina <jalil.salame@tum.de>