fix: Use VsockAddr instead of cid and port where appropiate
This follows the API of tokio where `bind` and `connect` take an `addr`
argument instead of a separate IP/hostname and port.

Tokio actually takes a `addr: impl ToSockAddrs` because of DNS
resolution (ie. `localhost` can be both `127.0.0.1` and `::1` or any
number of IPs for that matter). Vsocks do not do this so it makes sense
to keep the argument the same (although we could take `addr: impl
Into<Vsock>` instead in case we add some interesting `From` and
`TryFrom` methods).

Signed-off-by: Jalil David Salamé Messina <jalil.salame@gmail.com>